moved dao/model for user module
