+++ frontapi/src/main/java/com/harmoni/frontapi/main/user/dao/DevUserDaoService.java
@@ -1,6 +1,6 @@
-package com.harmoni.frontapi.main.user.service.dao;
+package com.harmoni.frontapi.main.user.dao;
 
-import com.harmoni.frontapi.main.user.service.model.User;
+import com.harmoni.frontapi.main.user.model.User;
 import jakarta.annotation.Nullable;
 import org.springframework.stereotype.Repository;
