+++ frontapi/src/main/java/com/harmoni/frontapi/main/user/model/Role.java
@@ -1,4 +1,4 @@
-package com.harmoni.frontapi.main.user.service.model;
+package com.harmoni.frontapi.main.user.model;
 
 public enum Role {
     ADMIN,
